bdev/nvme: Add dynamic allocation of namespaces
bdev_nvme layer allocates memory for maximum allowed number of
namespaces reported by controller. When number of allowed namespaces
is large and most of them are inactive it is waist of memory.

This patch changes allocation and storing method for
namespaces. Memory is allocated only for active namespaces and they
are stored in linked list. Access by namespace ID is done via linear
search. Since we do not access this array in data path, access speed
is not that important. It can be optimized later if needed.

Signed-off-by: Evgeniy Kochetov <evgeniik@nvidia.com>